Introduce Typescript
This is an initial minimal conversion.
More code will be converted later.
The focus has been on type the public API in `index`.
