+++ .eslintrc.json
@@ -6,10 +6,38 @@
     ],
     "env": {
         "node": true,
-        "amd": true,
+        //"amd": true,
         "es6": true
     },
-    "ignorePatterns": [ "./node_modules/**" ],
+    "parserOptions": {
+        "sourceType": "module"
+    },
+    "ignorePatterns": [ "/dist/**" ],
+    "overrides": [
+        {
+            "files": ["**/*.ts"],
+            "extends": [
+                "eslint:recommended",
+                "plugin:mocha/recommended",
+                "plugin:@typescript-eslint/eslint-recommended",
+                "plugin:@typescript-eslint/recommended"
+            ],
+            "parser":  "@typescript-eslint/parser",
+            "parserOptions": {
+                "project": "./tsconfig.json"
+            },
+            "plugins": [
+                "@typescript-eslint",
+                "mocha"
+            ]
+        },
+        {
+            "files": ["test/*.js"],
+            "env": {
+                "mocha": true
+            }
+        }
+    ],
     "plugins": [
         "mocha"
     ],
