+++ .gitignore
@@ -3,5 +3,6 @@
 *.jpg
 *.png
 node_modules
+dist
 package-lock.json
 .idea
