+++ package.json
@@ -3,7 +3,8 @@
   "version": "0.2.5",
   "description": "Pure JavaScript ID3v2 Tag writer and reader",
   "author": "Jan Metzger <jan.metzger@gmx.net>",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Zazama/node-id3.git"
@@ -28,26 +29,36 @@
   ],
   "homepage": "https://github.com/Zazama/node-id3#readme",
   "directories": {
-    "test": "test",
-    "example": "example"
+    "test": "dist/test",
+    "example": "dist/example"
   },
   "scripts": {
-    "test": "mocha",
+    "build": "tsc --build",
+    "clean": "tsc --build --clean",
+    "test": "mocha --require source-map-support/register dist/test",
+    "prepublish": "tsc",
+    "markdownlint-ci": "markdownlint README.md CHANGELOG.md",
     "markdownlint": "markdownlint --fix README.md CHANGELOG.md",
-    "eslint": "eslint **/*.js",
-    "eslint-fix": "eslint --fix **/*.js"
+    "eslint": "eslint **/*.js **/*.ts",
+    "eslint-fix": "eslint --fix **/*.js **/*.ts"
   },
   "license": "MIT",
   "dependencies": {
     "iconv-lite": "0.6.2"
   },
   "devDependencies": {
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.11.10",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "eslint": "^8.28.0",
     "eslint-plugin-mocha": "^10.1.0",
     "jsmediatags": "^3.9.3",
     "markdownlint-cli": "^0.32.2",
-    "mocha": "8.1.3"
+    "mocha": "8.1.3",
+    "source-map-support": "^0.5.21",
+    "typescript": "^4.9.3"
   }
 }
