+++ test/small-image.ts
@@ -1,0 +1,16 @@
+
+/**
+ * A small PNG image of 1x1 size with a white pixel.
+ */
+export const smallPngImageData = Buffer.from(
+    // 32 values per line.
+    "89504E470D0A1A0A0000000D4948445200000001000000010802000000907753"+
+    "DE000000017352474200AECE1CE90000000467414D410000B18F0BFC61050000"+
+    "00097048597300000EC300000EC301C76FA8640000000C49444154185763F8FF"+
+    "FF3F0005FE02FEA73581840000000049454E44AE426082",
+    "hex"
+)
+
+export const smallJpegImageData = Buffer.from(
+    '5B307836312C20307836322C20307836332C20307836345D', 'hex'
+)
